moved _fast_kde_2d function to plot_utils.py
